+++ packages/cli/uni-builder/src/types.ts
@@ -12,6 +12,8 @@ import type {
   RsbuildEntry,
   PromiseOrNot,
   RsbuildPluginAPI,
+  ArrayOrNot,
+  HtmlTagDescriptor,
 } from '@rsbuild/shared';
 import type { RsbuildConfig } from '@rsbuild/core';
 import type { PluginAssetsRetryOptions } from '@rsbuild/plugin-assets-retry';
@@ -31,6 +33,7 @@ import type {
   UniBuilderStartServerResult,
 } from './shared/devServer';
 import type { PluginSourceBuildOptions } from '@rsbuild/plugin-source-build';
+import type TerserPlugin from 'terser-webpack-plugin';
 
 export type CreateBuilderCommonOptions = {
   entry?: RsbuildEntry;
@@ -85,6 +88,11 @@ export type ToolsDevServerConfig = ChainedConfig<{
   proxy?: ServerConfig['proxy'];
 }>;
 
+export type TerserPluginOptions = TerserPlugin.BasePluginOptions &
+  TerserPlugin.DefinedDefaultMinimizerAndOptions<TerserPlugin.TerserOptions>;
+
+export type ToolsTerserConfig = ChainedConfig<TerserPluginOptions>;
+
 export type UniBuilderExtraConfig = {
   tools?: {
     styledComponents?: false | PluginStyledComponentsOptions;
@@ -115,6 +123,11 @@ export type UniBuilderExtraConfig = {
      * Tips: this configuration is not yet supported in rspack
      */
     tsLoader?: PluginTsLoaderOptions;
+    /**
+     * Modify the options of [terser-webpack-plugin](https://github.com/webpack-contrib/terser-webpack-plugin).
+     * @requires webpack
+     */
+    terser?: ToolsTerserConfig;
   };
   dev?: {
     /**
@@ -153,6 +166,11 @@ export type UniBuilderExtraConfig = {
     resolveExtensionPrefix?: string | Partial<Record<RsbuildTarget, string>>;
   };
   output?: {
+    /**
+     * Whether to disable code minification in production build.
+     */
+    // TODO: support output.minify configuration
+    disableMinimize?: boolean;
     /**
      * @deprecated use `output.filenameHash` instead
      */
@@ -225,6 +243,10 @@ export type UniBuilderExtraConfig = {
      * When this option is enabled, the generated HTML file path will change from `[name]/index.html` to `[name].html`.
      */
     disableHtmlFolder?: boolean;
+    /**
+     *  @deprecated use `html.tags` instead
+     */
+    tagsByEntries?: Record<string, ArrayOrNot<HtmlTagDescriptor>>;
     /**
      * @deprecated use `html.meta` instead
      */