+++ packages/generator/generators/tailwindcss-generator/templates/ts-template/tailwind.config.ts
@@ -2,4 +2,4 @@ import type { Config } from 'tailwindcss';
 
 export default {
   content: ['./src/**/*.{js,jsx,ts,tsx}'],
-} as Config;
+} satisfies Config;