+++ packages/devtools/plugin/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @modern-js/plugin-devtools
 
+## 2.48.3
+
+### Patch Changes
+
+- @modern-js/devtools-kit@2.48.3
+- @modern-js/devtools-client@2.48.3
+- @modern-js/runtime@2.48.3
+- @modern-js/utils@2.48.3
+
 ## 2.48.2
 
 ### Patch Changes