+++ packages/server/create-request/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @modern-js/create-request
 
+## 2.48.3
+
+### Patch Changes
+
+- @modern-js/runtime-utils@2.48.3
+- @modern-js/utils@2.48.3
+
 ## 2.48.2
 
 ### Patch Changes